refactor some part of helpers and split into related packages
