+++ mesheryctl/pkg/utils/auth.go
@@ -1,0 +1,77 @@
+package utils
+
+import (
+	"bytes"
+	"encoding/json"
+	"io/ioutil"
+	"net/http"
+	"os"
+
+	"github.com/layer5io/meshery/mesheryctl/internal/cli/root/config"
+	"github.com/pkg/errors"
+	"github.com/spf13/viper"
+)
+
+// AddAuthDetails Adds authentication cookies to the request
+func AddAuthDetails(req *http.Request, filepath string) error {
+	file, err := ioutil.ReadFile(filepath)
+	if err != nil {
+		err = errors.Wrap(err, "could not read token:")
+		return err
+	}
+	var tokenObj map[string]string
+	if err := json.Unmarshal(file, &tokenObj); err != nil {
+		err = errors.Wrap(err, "token file invalid:")
+		return err
+	}
+	req.AddCookie(&http.Cookie{
+		Name:     tokenName,
+		Value:    tokenObj[tokenName],
+		HttpOnly: true,
+	})
+	req.AddCookie(&http.Cookie{
+		Name:     providerName,
+		Value:    tokenObj[providerName],
+		HttpOnly: true,
+	})
+	return nil
+}
+
+// UpdateAuthDetails checks gets the token (old/refreshed) from meshery server and writes it back to the config file
+func UpdateAuthDetails(filepath string) error {
+	mctlCfg, err := config.GetMesheryCtl(viper.GetViper())
+	if err != nil {
+		return errors.Wrap(err, "error processing config")
+	}
+
+	// TODO: get this from the global config
+	req, err := http.NewRequest("GET", mctlCfg.GetBaseMesheryURL()+"/api/gettoken", bytes.NewBuffer([]byte("")))
+	if err != nil {
+		err = errors.Wrap(err, "error Creating the request :")
+		return err
+	}
+	if err := AddAuthDetails(req, filepath); err != nil {
+		return err
+	}
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	defer SafeClose(resp.Body)
+
+	if err != nil {
+		err = errors.Wrap(err, "error dispatching there request :")
+		return err
+	}
+
+	data, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		err = errors.Wrap(err, "error reading body :")
+		return err
+	}
+
+	if ContentTypeIsHTML(resp) {
+		return errors.New("invalid body")
+	}
+
+	return ioutil.WriteFile(filepath, data, os.ModePerm)
+}
