+++ mesheryctl/pkg/utils/healthcheck.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"context"
 	"fmt"
 	"os/exec"
 	"regexp"
@@ -8,6 +9,8 @@ import (
 	"strings"
 
 	meshkitkube "github.com/layer5io/meshkit/utils/kubernetes"
+	"github.com/pkg/errors"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/version"
 )
 
@@ -107,4 +110,50 @@ var semVer = regexp.MustCompile("(v[0-9]+.[0-9]+.[0-9]+)")
 func parseKubectlShortVersion(version string) ([3]int, error) {
 	versionString := semVer.FindString(version)
 	return getK8sVersion(versionString)
+}
+
+// IsMesheryRunning checks if the meshery server containers are up and running
+func IsMesheryRunning(currPlatform string) (bool, error) {
+	switch currPlatform {
+	case "docker":
+		{
+			op, err := exec.Command("docker-compose", "-f", DockerComposeFile, "ps").Output()
+			if err != nil {
+				return false, err
+			}
+			return strings.Contains(string(op), "meshery"), nil
+		}
+	case "kubernetes":
+		{
+			client, err := meshkitkube.New([]byte(""))
+
+			if err != nil {
+				return false, errors.Wrap(err, "failed to create new client")
+			}
+
+			//podInterface := client.KubeClient.CoreV1().Pods(MesheryNamespace)
+			deploymentInterface := client.KubeClient.AppsV1().Deployments(MesheryNamespace)
+			//podList, err := podInterface.List(context.TODO(), v1.ListOptions{})
+			deploymentList, err := deploymentInterface.List(context.TODO(), v1.ListOptions{})
+
+			if err != nil {
+				return false, err
+			}
+			//for i, pod := range podList.Items {
+			//	fmt.Println(i, pod.GetName())
+			//	//if strings.Contains(pod.GetName(), "meshery") {
+			//	//	return true, nil
+			//	//}
+			//}
+			for _, deployment := range deploymentList.Items {
+				if deployment.GetName() == "meshery" {
+					return true, nil
+				}
+			}
+
+			return false, err
+		}
+	}
+
+	return false, nil
 }
